+++ src/features/token-migration/components/migration-details/__tests__/migration-details.test.tsx
@@ -4,15 +4,16 @@ import { screen } from "@testing-library/react"
 import { useParams } from "react-router-dom"
 import { useSingleTransactionList } from "features/transactions/queries"
 import { useGetBlock } from "features/network/queries"
-import { createMockSendTxList, mockSingleTxListError } from "test/transactions"
 import { mockBlockError } from "test/block"
 import {
+  createMockSendTxList,
   mockBlockHeight,
   mockHash,
   mockInvalidEventId,
   mockLongBlockHeight,
   mockLongEventId,
   mockSendEventId,
+  mockSingleTxListError,
   mockUseBlock,
   mockUseParams,
   mockUserAddr,
